packages/sanjaya-ninja/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sanjaya-ninja"
version = "1.0.0"
description = "Django Ninja app for dynamic reporting with pluggable data providers"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    { name = "Tom Brennan" },
]
keywords = ["django", "reporting", "analytics", "ag-grid", "pivot", "ninja"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Typing :: Typed",
]
dependencies = [
    "django>=4.2,<6",
    "django-ninja>=1.0,<2",
    "sanjaya-core~=1.0",
    "openpyxl>=3.1,<4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-django>=4.5",
    "pytest-cov",
]

[project.urls]
Repository = "https://github.com/tjb1982/sanjaya"
Issues = "https://github.com/tjb1982/sanjaya/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/sanjaya_ninja"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
DJANGO_SETTINGS_MODULE = "tests.settings"
